@@ -12,9 +12,9 @@
         }
     ],
     "require": {
-        "php": ">=5.6.0",
+        "php": "^7.0",
         "psr/log": "~1.0",
-        "elasticsearch/elasticsearch": "5.3.0"
+        "elasticsearch/elasticsearch": "^6.0"
     },
     "require-dev": {
         "aws/aws-sdk-php": "~3.0",
@@ -38,7 +38,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.2.x-dev"
+            "dev-master": "6.0.x-dev"
         }
     }
 }
