@@ -1,5 +1,11 @@
 <?php
+
+#declare(strict_types = 1);
+
+namespace Elasticsearch\Tests\ConnectionPool;
+
 use Elasticsearch\ClientBuilder;
+use Elasticsearch\ConnectionPool\SniffingConnectionPool;
 
 /**
  * Class SniffingConnectionPoolIntegrationTest
@@ -11,15 +17,16 @@ use Elasticsearch\ClientBuilder;
  * @license    http://www.apache.org/licenses/LICENSE-2.0 Apache2
  * @link       http://elasticsearch.org
  */
-class SniffingConnectionPoolIntegrationTest extends \PHPUnit_Framework_TestCase
+class SniffingConnectionPoolIntegrationTest extends \PHPUnit\Framework\TestCase
 {
     public function testSniff()
     {
         $client = ClientBuilder::create()
             ->setHosts([$_SERVER['ES_TEST_HOST']])
-            ->setConnectionPool('\Elasticsearch\ConnectionPool\SniffingConnectionPool', ['sniffingInterval' => -10])
+            ->setConnectionPool(SniffingConnectionPool::class, ['sniffingInterval' => -10])
             ->build();
 
-        $client->ping();
+        $pinged = $client->ping();
+        $this->assertTrue($pinged);
     }
 }
