@@ -1,7 +1,10 @@
 <?php
 
+#declare(strict_types = 1);
+
 namespace Elasticsearch\Tests\Helper\Iterators;
 
+use Elasticsearch\Client;
 use Elasticsearch\Helper\Iterators\SearchResponseIterator;
 use Mockery as m;
 
@@ -12,7 +15,7 @@ use Mockery as m;
  * @license http://www.apache.org/licenses/LICENSE-2.0 Apache2
  * @link    http://Elasticsearch.org
  */
-class SearchResponseIteratorTest extends \PHPUnit_Framework_TestCase
+class SearchResponseIteratorTest extends \PHPUnit\Framework\TestCase
 {
 
     public function tearDown()
@@ -22,24 +25,24 @@ class SearchResponseIteratorTest extends \PHPUnit_Framework_TestCase
 
     public function testWithNoResults()
     {
-        $search_params = array(
+        $search_params = [
             'scroll'      => '5m',
             'index'       => 'twitter',
             'size'        => 1000,
-            'body'        => array(
-                'query' => array(
+            'body'        => [
+                'query' => [
                     'match_all' => new \StdClass
-                )
-            )
-        );
+                ]
+            ]
+        ];
 
-        $mock_client = m::mock('\Elasticsearch\Client');
+        $mock_client = m::mock(Client::class);
 
         $mock_client->shouldReceive('search')
             ->once()
             ->ordered()
             ->with($search_params)
-            ->andReturn(array('_scroll_id' => 'scroll_id_01'));
+            ->andReturn(['_scroll_id' => 'scroll_id_01']);
 
         $mock_client->shouldReceive('scroll')
             ->never();
@@ -57,18 +60,18 @@ class SearchResponseIteratorTest extends \PHPUnit_Framework_TestCase
 
     public function testWithHits()
     {
-        $search_params = array(
+        $search_params = [
             'scroll'      => '5m',
             'index'       => 'twitter',
             'size'        => 1000,
-            'body'        => array(
-                'query' => array(
+            'body'        => [
+                'query' => [
                     'match_all' => new \StdClass
-                )
-            )
-        );
+                ]
+            ]
+        ];
 
-        $mock_client = m::mock('\Elasticsearch\Client');
+        $mock_client = m::mock(Client::class);
 
         $mock_client->shouldReceive('search')
             ->once()
@@ -104,7 +107,8 @@ class SearchResponseIteratorTest extends \PHPUnit_Framework_TestCase
                             ]
                         ]
                     ]
-                ]);
+                ]
+            );
 
         $mock_client->shouldReceive('scroll')
             ->once()
