@@ -1,5 +1,7 @@
 <?php
 
+#declare(strict_types = 1);
+
 namespace Elasticsearch\Tests;
 
 use Doctrine\Common\Inflector\Inflector;
@@ -11,12 +13,13 @@ use Elasticsearch\Common\Exceptions\Missing404Exception;
 use Elasticsearch\Common\Exceptions\RequestTimeout408Exception;
 use Elasticsearch\Common\Exceptions\ServerErrorResponseException;
 use Elasticsearch\Common\Exceptions\RoutingMissingException;
+use Elasticsearch\Common\Exceptions\Unauthorized401Exception;
 use GuzzleHttp\Ring\Future\FutureArrayInterface;
-use stdClass;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
 use Symfony\Component\Yaml\Exception\ParseException;
-use Symfony\Component\Yaml\Parser;
 use Symfony\Component\Yaml\Yaml;
 
 /**
@@ -29,9 +32,9 @@ use Symfony\Component\Yaml\Yaml;
  * @license    http://www.apache.org/licenses/LICENSE-2.0 Apache2
  * @link       http://elasticsearch.org
  */
-class YamlRunnerTest extends \PHPUnit_Framework_TestCase
+class YamlRunnerTest extends \PHPUnit\Framework\TestCase
 {
-    /** @var Parser Yaml parser for reading integrations tests */
+    /** @var \Symfony\Component\Yaml\Yaml Yaml parser for reading integrations tests */
     private $yaml;
 
     /** @var Elasticsearch\Client client used by elasticsearch */
@@ -42,7 +45,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
 
     /** @var array A list of supported features */
     private static $supportedFeatures = [
-        'stash_in_path', 'warnings'
+        'stash_in_path', 'warnings', 'headers'
     ];
 
     /** @var array A mapping for endpoint when there is a reserved keywords for the method / namespace name */
@@ -52,20 +55,42 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         ],
     ];
 
+    private static $skippedTests = [
+
+    ];
+
     /** @var array A list of skipped test with their reasons */
-    private static $skippedTest = [
+    private static $skippedFiles = [
+
+        'cat.nodeattrs/10_basic.yml' => 'Using java regex fails in PHP',
         'cat.nodeattrs/10_basic.yaml' => 'Using java regex fails in PHP',
+
+        'cat.repositories/10_basic.yml' => 'Using java regex fails in PHP',
         'cat.repositories/10_basic.yaml' => 'Using java regex fails in PHP',
+
+        'indices.shrink/10_basic.yml' => 'Shrink tests seem to require multiple nodes',
         'indices.shrink/10_basic.yaml' => 'Shrink tests seem to require multiple nodes',
-        'indices.rollover/10_basic.yaml' => 'Rollover test seems buggy atm'
+
+        'indices.rollover/10_basic.yml' => 'Rollover test seems buggy atm',
+        'indices.rollover/10_basic.yaml' => 'Rollover test seems buggy atm',
+
     ];
 
     /** @var array A list of files to skip completely, due to fatal parsing errors */
-    private static $skippedFiles = [
+    private static $fatalFiles = [
+        'indices.create/10_basic.yml' => 'Temporary: Yaml parser doesnt support "inline" empty keys',
         'indices.create/10_basic.yaml' => 'Temporary: Yaml parser doesnt support "inline" empty keys',
+
+        'indices.put_mapping/10_basic.yml' => 'Temporary: Yaml parser doesnt support "inline" empty keys',
         'indices.put_mapping/10_basic.yaml' => 'Temporary: Yaml parser doesnt support "inline" empty keys',
+
+        'search/110_field_collapsing.yml' => 'Temporary: parse error, malformed inline yaml',
         'search/110_field_collapsing.yaml' => 'Temporary: parse error, malformed inline yaml',
-        'cat.nodes/10_basic.yaml' => 'Temporary: parse error, something about $body: |'
+
+        'cat.nodes/10_basic.yml' => 'Temporary: parse error, something about $body: |',
+        'cat.nodes/10_basic.yaml' => 'Temporary: parse error, something about $body: |',
+        'search.aggregation/180_percentiles_tdigest_metric.yml' => 'array of objects, unclear how to fix',
+        'search.aggregation/190_percentiles_hdr_metric.yml' => 'array of objects, unclear how to fix'
     ];
 
     /**
@@ -105,11 +130,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
     public function setUp()
     {
         $this->clean();
-        $builder = Elasticsearch\ClientBuilder::create()->setHosts([self::getHost()]);
-        if (version_compare(phpversion(), '5.6.6', '<') || ! defined('JSON_PRESERVE_ZERO_FRACTION')) {
-            $builder->allowBadJSONSerialization();
-        }
-        $this->client = $builder->build();
+        $this->client = Elasticsearch\ClientBuilder::create()->setHosts([self::getHost()])->build();
     }
 
     /**
@@ -122,16 +143,16 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
             static::markTestIncomplete($testProcedure);
         }
 
-        if (array_key_exists($fileName, static::$skippedTest)) {
-            static::markTestSkipped(static::$skippedTest[$fileName]);
+        if (array_key_exists($fileName, static::$skippedFiles)) {
+            static::markTestSkipped(static::$skippedFiles[$fileName]);
         }
 
         if (null !== $setupProcedure) {
-            $this->processProcedure(current($setupProcedure), 'setup');
+            $this->processProcedure(current($setupProcedure), 'setup', $fileName);
             $this->waitForYellow();
         }
 
-        $this->processProcedure(current($testProcedure), key($testProcedure));
+        $this->processProcedure(current($testProcedure), key($testProcedure), $fileName);
     }
 
     /**
@@ -144,16 +165,16 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
             static::markTestIncomplete($testProcedure);
         }
 
-        if (array_key_exists($fileName, static::$skippedTest)) {
-            static::markTestSkipped(static::$skippedTest[$fileName]);
+        if (array_key_exists($fileName, static::$skippedFiles)) {
+            static::markTestSkipped(static::$skippedFiles[$fileName]);
         }
 
         if (null !== $setupProcedure) {
-            $this->processProcedure(current($setupProcedure), 'setup');
+            $this->processProcedure(current($setupProcedure), 'setup', $fileName);
             $this->waitForYellow();
         }
 
-        $this->processProcedure(current($testProcedure), key($testProcedure), true);
+        $this->processProcedure(current($testProcedure), key($testProcedure), $fileName, true);
     }
 
     /**
@@ -163,10 +184,14 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
      * @param $name
      * @param bool $async
      */
-    public function processProcedure($procedure, $name, $async = false)
+    public function processProcedure($procedure, $name, $fileName, $async = false)
     {
         $lastOperationResult = null;
         $context = [];
+
+        if (array_key_exists("$fileName#$name", static::$skippedTests)) {
+            static::markTestSkipped(static::$skippedTests["$fileName#$name"]);
+        }
 
         foreach ($procedure as $operation) {
             $lastOperationResult = $this->processOperation($operation, $lastOperationResult, $context, $name, $async);
@@ -179,7 +204,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
      * @param      $operation
      * @param      $lastOperationResult
      * @param      $testName
-     * @param array $context 
+     * @param array $context
      * @param bool $async
      *
      * @return mixed
@@ -252,6 +277,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
     {
         $expectedError = null;
         $expectedWarnings = null;
+        $headers = null;
 
         // Check if a error must be caught
         if ('catch' === key($operation)) {
@@ -262,6 +288,12 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         // Check if a warning must be caught
         if ('warnings' === key($operation)) {
             $expectedWarnings = current($operation);
+            next($operation);
+        }
+
+        // Any specific headers to add?
+        if ('headers' === key($operation)) {
+            $headers = current($operation);
             next($operation);
         }
 
@@ -291,6 +323,13 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
             $endpointParams->client['future'] = true;
         }
 
+        if ($headers != null) {
+            $endpointParams->client['headers'] = $headers;
+        }
+
+        if (!is_string($method)) {
+            throw new \Exception('$method must be string');
+        }
         list($method, $namespace) = $this->mapEndpoint($method, $namespace);
 
         if (null !== $namespace) {
@@ -307,7 +346,9 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
 
         // TODO remove this after cat testing situation resolved
         if ($caller instanceof Elasticsearch\Namespaces\CatNamespace) {
-            $endpointParams->format = 'text';
+            if (!isset($endpointParams->format)) {
+                $endpointParams->format = 'text';
+            }
         }
 
         // Exist* methods have to be manually 'unwrapped' into true/false for async
@@ -367,12 +408,11 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
      *
      * @throws \Exception
      *
-     * @return bool
+     * @return bool|mixed[]
      */
     public function executeAsyncExistRequest($caller, $method, $endpointParams, $expectedError, $expectedWarnings, $testName)
     {
         try {
-
             $response = $caller->$method($endpointParams);
 
             while ($response instanceof FutureArrayInterface) {
@@ -399,7 +439,8 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         }
     }
 
-    public function checkForWarnings($expectedWarnings) {
+    public function checkForWarnings($expectedWarnings)
+    {
         $last = $this->client->transport->getLastConnection()->getLastRequestInfo();
 
 
@@ -431,11 +472,10 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         }
 
         // Check to make sure we're adding headers
-        static::assertArrayHasKey('Content-type', $last['request']['headers'], print_r($last['request']['headers'], true));
-        static::assertEquals('application/json', $last['request']['headers']['Content-type'][0], print_r($last['request']['headers'], true));
-        static::assertArrayHasKey('Accept', $last['request']['headers'], print_r($last['request']['headers'], true));
-        static::assertEquals('application/json', $last['request']['headers']['Accept'][0], print_r($last['request']['headers'], true));
-
+        $this->assertArrayHasKey('Content-Type', $last['request']['headers'], print_r($last['request']['headers'], true));
+        $this->assertSame('application/json', $last['request']['headers']['Content-Type'][0], print_r($last['request']['headers'], true));
+        $this->assertArrayHasKey('Accept', $last['request']['headers'], print_r($last['request']['headers'], true));
+        $this->assertSame('application/json', $last['request']['headers']['Accept'][0], print_r($last['request']['headers'], true));
     }
 
     /**
@@ -451,7 +491,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $msg = "Failed to assert that a value is false in test \"$testName\"\n"
             ."$operation was [".print_r($value, true)."]"
             .var_export($lastOperationResult, true);
-        static::assertFalse($value, $msg);
+        $this->assertFalse($value, $msg);
 
         return $lastOperationResult;
     }
@@ -470,10 +510,10 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $msg = "Failed to assert that a value is true in test \"$testName\"\n"
             ."$operation was [".print_r($value, true)."]"
             .var_export($lastOperationResult, true);
-        static::assertNotEquals(0, $value, $msg);
-        static::assertNotFalse($value, $msg);
-        static::assertNotNull($value, $msg);
-        static::assertNotEquals('', $msg);
+        $this->assertNotEquals(0, $value, $msg);
+        $this->assertNotFalse($value, $msg);
+        $this->assertNotNull($value, $msg);
+        $this->assertNotEquals('', $msg);
 
         return $lastOperationResult;
     }
@@ -504,11 +544,11 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
             $expected = json_decode(json_encode($expected), true);
             $match = json_decode(json_encode($match), true);
 
-            static::assertEquals($expected, $match, $msg);
+            $this->assertEquals($expected, $match, $msg);
         } elseif (is_string($expected) && preg_match('#^/.+?/$#s', $expected)) {
-            static::assertRegExp($this->formatRegex($expected), $match, $msg);
+            $this->assertRegExp($this->formatRegex($expected), $match, $msg);
         } else {
-            static::assertEquals($expected, $match, $msg);
+            $this->assertEquals($expected, $match, $msg);
         }
 
         return $lastOperationResult;
@@ -526,7 +566,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $value = $this->resolveValue($lastOperationResult, key($operation), $context);
         $expected = current($operation);
 
-        static::assertGreaterThanOrEqual($expected, $value, 'Failed to gte in test ' . $testName);
+        $this->assertGreaterThanOrEqual($expected, $value, 'Failed to gte in test ' . $testName);
 
         return $lastOperationResult;
     }
@@ -543,7 +583,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $value = $this->resolveValue($lastOperationResult, key($operation), $context);
         $expected = current($operation);
 
-        static::assertGreaterThan($expected, $value, 'Failed to gt in test ' . $testName);
+        $this->assertGreaterThan($expected, $value, 'Failed to gt in test ' . $testName);
 
         return $lastOperationResult;
     }
@@ -560,7 +600,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $value = $this->resolveValue($lastOperationResult, key($operation), $context);
         $expected = current($operation);
 
-        static::assertLessThanOrEqual($expected, $value, 'Failed to lte in test ' . $testName);
+        $this->assertLessThanOrEqual($expected, $value, 'Failed to lte in test ' . $testName);
 
         return $lastOperationResult;
     }
@@ -577,7 +617,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $value = $this->resolveValue($lastOperationResult, key($operation), $context);
         $expected = current($operation);
 
-        static::assertLessThan($expected, $value, 'Failed to lt in test ' . $testName);
+        $this->assertLessThan($expected, $value, 'Failed to lt in test ' . $testName);
 
         return $lastOperationResult;
     }
@@ -594,7 +634,7 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $value = $this->resolveValue($lastOperationResult, key($operation), $context);
         $expected = current($operation);
 
-        static::assertCount($expected, $value, 'Failed to gte in test ' . $testName);
+        $this->assertCount($expected, $value, 'Failed to gte in test ' . $testName);
 
         return $lastOperationResult;
     }
@@ -627,12 +667,20 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
      */
     public function operationSkip($operation, $lastOperationResult, $testName)
     {
-        if (is_object($operation) !== true ) {
+        if (is_object($operation) !== true) {
             return $lastOperationResult;
         }
 
-        if (property_exists($operation, 'features') && !in_array($operation->features, static::$supportedFeatures, true)) {
-            static::markTestSkipped(sprintf('Feature(s) %s not supported in test "%s"', json_encode($operation->features), $testName));
+        if (property_exists($operation, 'features')) {
+            if (is_array($operation->features)) {
+                if (count(array_intersect($operation->features, static::$supportedFeatures)) != count($operation->features)) {
+                    static::markTestSkipped(sprintf('Feature(s) %s not supported in test "%s"', json_encode($operation->features), $testName));
+                }
+            } else {
+                if (!in_array($operation->features, static::$supportedFeatures, true)) {
+                    static::markTestSkipped(sprintf('Feature(s) %s not supported in test "%s"', json_encode($operation->features), $testName));
+                }
+            }
         }
 
         if (property_exists($operation, 'version')) {
@@ -648,11 +696,11 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
                 $version[0] = ~PHP_INT_MAX;
             }
 
-            if (!isset($version[1]) || $version[1] === "" ) {
+            if (!isset($version[1]) || $version[1] === "") {
                 $version[1] = PHP_INT_MAX;
             }
 
-            if (version_compare(static::$esVersion, $version[0], '>=')  && version_compare(static::$esVersion, $version[1], '<=')) {
+            if (version_compare(static::$esVersion, (string)$version[0], '>=')  && version_compare(static::$esVersion, (string)$version[1], '<=')) {
                 static::markTestSkipped(sprintf('Skip test "%s", as version %s should be skipped (%s)', $testName, static::$esVersion, $operation->reason));
             }
         }
@@ -672,23 +720,27 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
     private function assertException(\Exception $exception, $expectedError, $testName)
     {
         if (is_string($expectedError) && preg_match('#^/.+?/$#', $expectedError)) {
-            static::assertRegExp($expectedError, $exception->getMessage(), 'Failed to catch error in test ' . $testName);
+            $this->assertRegExp($expectedError, $exception->getMessage(), 'Failed to catch error in test ' . $testName);
+        } elseif ($exception instanceof BadRequest400Exception && $expectedError === 'bad_request') {
+            $this->assertTrue(true);
+        } elseif ($exception instanceof Unauthorized401Exception && $expectedError === 'unauthorized') {
+            $this->assertTrue(true);
         } elseif ($exception instanceof Missing404Exception && $expectedError === 'missing') {
-            static::assertTrue(true);
+            $this->assertTrue(true);
         } elseif ($exception instanceof Conflict409Exception && $expectedError === 'conflict') {
-            static::assertTrue(true);
+            $this->assertTrue(true);
         } elseif ($exception instanceof Forbidden403Exception && $expectedError === 'forbidden') {
-            static::assertTrue(true);
+            $this->assertTrue(true);
         } elseif ($exception instanceof RequestTimeout408Exception && $expectedError === 'request_timeout') {
-            static::assertTrue(true);
+            $this->assertTrue(true);
         } elseif ($exception instanceof BadRequest400Exception && $expectedError === 'request') {
-            static::assertTrue(true);
+            $this->assertTrue(true);
         } elseif ($exception instanceof ServerErrorResponseException && $expectedError === 'request') {
-            static::assertTrue(true);
+            $this->assertTrue(true);
         } elseif ($exception instanceof \RuntimeException && $expectedError === 'param') {
-            static::assertTrue(true);
+            $this->assertTrue(true);
         } else {
-            static::assertContains($expectedError, $exception->getMessage());
+            $this->assertContains($expectedError, $exception->getMessage());
         }
 
         if ($exception->getPrevious() !== null) {
@@ -712,6 +764,9 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $finder = new Finder();
         $finder->in($path);
         $finder->files();
+        $finder->name('*.yml');
+
+        // *.yaml files should be included until the library is ES 6.0+ only
         $finder->name('*.yaml');
 
         $filter = isset($_SERVER['TEST_CASE']) ? $_SERVER['TEST_CASE'] : null;
@@ -845,7 +900,11 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
      */
     private function splitDocument($file, $path, $filter = null)
     {
-        $fileContent = file_get_contents($file);
+
+        $fileContent = $file->getContents();
+        // cleanup some bad comments
+        $fileContent = str_replace('"#', '" #', $fileContent);
+
         $documents = explode("---\n", $fileContent);
         $documents = array_filter($documents, function ($item) {
             return trim($item) !== '';
@@ -856,8 +915,8 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         $setupSkip = false;
         $fileName = str_replace($path . '/', '', $file);
 
-        if (array_key_exists($fileName, static::$skippedFiles)) {
-            echo "Skipping: $fileName.  ".static::$skippedFiles[$fileName]."\n";
+        if (array_key_exists($fileName, static::$fatalFiles)) {
+            echo "Skipping: $fileName.  ".static::$fatalFiles[$fileName]."\n";
             return [];
         }
 
@@ -974,19 +1033,35 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
 
                     $response = curl_exec($ch);
                     curl_close($ch);
-
                 }
             }
         }
 
-        // TODO ewwww...
-        shell_exec('rm -rf /tmp/test_repo_create_1_loc');
-        shell_exec('rm -rf /tmp/test_repo_restore_1_loc');
-        shell_exec('rm -rf /tmp/test_cat_repo_1_loc');
-        shell_exec('rm -rf /tmp/test_cat_repo_2_loc');
-        shell_exec('rm -rf /tmp/test_cat_snapshots_1_loc');
+        $this->rmDirRecursively('/tmp/test_repo_create_1_loc');
+        $this->rmDirRecursively('/tmp/test_repo_restore_1_loc');
+        $this->rmDirRecursively('/tmp/test_cat_repo_1_loc');
+        $this->rmDirRecursively('/tmp/test_cat_repo_2_loc');
+        $this->rmDirRecursively('/tmp/test_cat_snapshots_1_loc');
 
         $this->waitForYellow();
+    }
+
+    private function rmDirRecursively($dir)
+    {
+        if (!is_dir($dir)) {
+            return;
+        }
+        $files = new RecursiveIteratorIterator(
+            new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS),
+            RecursiveIteratorIterator::CHILD_FIRST
+        );
+
+        foreach ($files as $fileinfo) {
+            $todo = ($fileinfo->isDir() ? 'rmdir' : 'unlink');
+            $todo($fileinfo->getRealPath());
+        }
+
+        rmdir($dir);
     }
 
     /**
@@ -1009,8 +1084,8 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
             ++$counter;
 
             if ($counter > 10) {
-                $this->log("Aborting test due to failure in clearing cluster.\n");
-                $this->log(print_r($response, true));
+                echo "Aborting test due to failure in clearing cluster.\n";
+                echo print_r($response, true);
                 exit;
             }
         }
