@@ -148,6 +148,7 @@ return array(
     'Elastica\\Aggregation\\AbstractSimpleAggregation' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/AbstractSimpleAggregation.php',
     'Elastica\\Aggregation\\AbstractTermsAggregation' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/AbstractTermsAggregation.php',
     'Elastica\\Aggregation\\Avg' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/Avg.php',
+    'Elastica\\Aggregation\\AvgBucket' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/AvgBucket.php',
     'Elastica\\Aggregation\\BucketScript' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/BucketScript.php',
     'Elastica\\Aggregation\\Cardinality' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/Cardinality.php',
     'Elastica\\Aggregation\\Children' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/Children.php',
@@ -174,7 +175,9 @@ return array(
     'Elastica\\Aggregation\\SerialDiff' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/SerialDiff.php',
     'Elastica\\Aggregation\\SignificantTerms' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/SignificantTerms.php',
     'Elastica\\Aggregation\\Stats' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/Stats.php',
+    'Elastica\\Aggregation\\StatsBucket' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/StatsBucket.php',
     'Elastica\\Aggregation\\Sum' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/Sum.php',
+    'Elastica\\Aggregation\\SumBucket' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/SumBucket.php',
     'Elastica\\Aggregation\\Terms' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/Terms.php',
     'Elastica\\Aggregation\\TopHits' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/TopHits.php',
     'Elastica\\Aggregation\\ValueCount' => $vendorDir . '/ruflin/elastica/lib/Elastica/Aggregation/ValueCount.php',
@@ -235,6 +238,26 @@ return array(
     'Elastica\\Node\\Info' => $vendorDir . '/ruflin/elastica/lib/Elastica/Node/Info.php',
     'Elastica\\Node\\Stats' => $vendorDir . '/ruflin/elastica/lib/Elastica/Node/Stats.php',
     'Elastica\\Param' => $vendorDir . '/ruflin/elastica/lib/Elastica/Param.php',
+    'Elastica\\Pipeline' => $vendorDir . '/ruflin/elastica/lib/Elastica/Pipeline.php',
+    'Elastica\\Processor\\AbstractProcessor' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/AbstractProcessor.php',
+    'Elastica\\Processor\\Append' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Append.php',
+    'Elastica\\Processor\\Attachment' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Attachment.php',
+    'Elastica\\Processor\\Convert' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Convert.php',
+    'Elastica\\Processor\\Date' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Date.php',
+    'Elastica\\Processor\\DateIndexName' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/DateIndexName.php',
+    'Elastica\\Processor\\DotExpander' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/DotExpander.php',
+    'Elastica\\Processor\\Fail' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Fail.php',
+    'Elastica\\Processor\\Join' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Join.php',
+    'Elastica\\Processor\\Json' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Json.php',
+    'Elastica\\Processor\\Kv' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Kv.php',
+    'Elastica\\Processor\\Lowercase' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Lowercase.php',
+    'Elastica\\Processor\\Remove' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Remove.php',
+    'Elastica\\Processor\\Rename' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Rename.php',
+    'Elastica\\Processor\\Set' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Set.php',
+    'Elastica\\Processor\\Sort' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Sort.php',
+    'Elastica\\Processor\\Split' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Split.php',
+    'Elastica\\Processor\\Trim' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Trim.php',
+    'Elastica\\Processor\\Uppercase' => $vendorDir . '/ruflin/elastica/lib/Elastica/Processor/Uppercase.php',
     'Elastica\\Query' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query.php',
     'Elastica\\QueryBuilder' => $vendorDir . '/ruflin/elastica/lib/Elastica/QueryBuilder.php',
     'Elastica\\QueryBuilder\\DSL' => $vendorDir . '/ruflin/elastica/lib/Elastica/QueryBuilder/DSL.php',
@@ -249,7 +272,6 @@ return array(
     'Elastica\\Query\\AbstractGeoShape' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/AbstractGeoShape.php',
     'Elastica\\Query\\AbstractQuery' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/AbstractQuery.php',
     'Elastica\\Query\\AbstractSpanQuery' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/AbstractSpanQuery.php',
-    'Elastica\\Query\\Bool' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Bool.php',
     'Elastica\\Query\\BoolQuery' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/BoolQuery.php',
     'Elastica\\Query\\Boosting' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Boosting.php',
     'Elastica\\Query\\Common' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Common.php',
@@ -260,7 +282,6 @@ return array(
     'Elastica\\Query\\Fuzzy' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Fuzzy.php',
     'Elastica\\Query\\GeoBoundingBox' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/GeoBoundingBox.php',
     'Elastica\\Query\\GeoDistance' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/GeoDistance.php',
-    'Elastica\\Query\\GeoDistanceRange' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/GeoDistanceRange.php',
     'Elastica\\Query\\GeoPolygon' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/GeoPolygon.php',
     'Elastica\\Query\\GeoShapePreIndexed' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/GeoShapePreIndexed.php',
     'Elastica\\Query\\GeoShapeProvided' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/GeoShapeProvided.php',
@@ -268,7 +289,6 @@ return array(
     'Elastica\\Query\\HasChild' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/HasChild.php',
     'Elastica\\Query\\HasParent' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/HasParent.php',
     'Elastica\\Query\\Ids' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Ids.php',
-    'Elastica\\Query\\Indices' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Indices.php',
     'Elastica\\Query\\InnerHits' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/InnerHits.php',
     'Elastica\\Query\\Limit' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Limit.php',
     'Elastica\\Query\\Match' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Match.php',
@@ -288,11 +308,14 @@ return array(
     'Elastica\\Query\\Script' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Script.php',
     'Elastica\\Query\\Simple' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Simple.php',
     'Elastica\\Query\\SimpleQueryString' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SimpleQueryString.php',
+    'Elastica\\Query\\SpanContaining' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanContaining.php',
     'Elastica\\Query\\SpanFirst' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanFirst.php',
     'Elastica\\Query\\SpanMulti' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanMulti.php',
     'Elastica\\Query\\SpanNear' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanNear.php',
+    'Elastica\\Query\\SpanNot' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanNot.php',
     'Elastica\\Query\\SpanOr' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanOr.php',
     'Elastica\\Query\\SpanTerm' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanTerm.php',
+    'Elastica\\Query\\SpanWithin' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/SpanWithin.php',
     'Elastica\\Query\\Term' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Term.php',
     'Elastica\\Query\\Terms' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Terms.php',
     'Elastica\\Query\\Type' => $vendorDir . '/ruflin/elastica/lib/Elastica/Query/Type.php',
@@ -312,7 +335,6 @@ return array(
     'Elastica\\Script\\AbstractScript' => $vendorDir . '/ruflin/elastica/lib/Elastica/Script/AbstractScript.php',
     'Elastica\\Script\\Script' => $vendorDir . '/ruflin/elastica/lib/Elastica/Script/Script.php',
     'Elastica\\Script\\ScriptFields' => $vendorDir . '/ruflin/elastica/lib/Elastica/Script/ScriptFields.php',
-    'Elastica\\Script\\ScriptFile' => $vendorDir . '/ruflin/elastica/lib/Elastica/Script/ScriptFile.php',
     'Elastica\\Script\\ScriptId' => $vendorDir . '/ruflin/elastica/lib/Elastica/Script/ScriptId.php',
     'Elastica\\Scroll' => $vendorDir . '/ruflin/elastica/lib/Elastica/Scroll.php',
     'Elastica\\Search' => $vendorDir . '/ruflin/elastica/lib/Elastica/Search.php',
@@ -326,14 +348,13 @@ return array(
     'Elastica\\Suggest\\Completion' => $vendorDir . '/ruflin/elastica/lib/Elastica/Suggest/Completion.php',
     'Elastica\\Suggest\\Phrase' => $vendorDir . '/ruflin/elastica/lib/Elastica/Suggest/Phrase.php',
     'Elastica\\Suggest\\Term' => $vendorDir . '/ruflin/elastica/lib/Elastica/Suggest/Term.php',
-    'Elastica\\Tool\\CrossIndex' => $vendorDir . '/ruflin/elastica/lib/Elastica/Tool/CrossIndex.php',
+    'Elastica\\Task' => $vendorDir . '/ruflin/elastica/lib/Elastica/Task.php',
     'Elastica\\Transport\\AbstractTransport' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/AbstractTransport.php',
     'Elastica\\Transport\\AwsAuthV4' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/AwsAuthV4.php',
     'Elastica\\Transport\\Guzzle' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/Guzzle.php',
     'Elastica\\Transport\\Http' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/Http.php',
     'Elastica\\Transport\\HttpAdapter' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/HttpAdapter.php',
     'Elastica\\Transport\\Https' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/Https.php',
-    'Elastica\\Transport\\Null' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/Null.php',
     'Elastica\\Transport\\NullTransport' => $vendorDir . '/ruflin/elastica/lib/Elastica/Transport/NullTransport.php',
     'Elastica\\Type' => $vendorDir . '/ruflin/elastica/lib/Elastica/Type.php',
     'Elastica\\Type\\AbstractType' => $vendorDir . '/ruflin/elastica/lib/Elastica/Type/AbstractType.php',
@@ -365,6 +386,7 @@ return array(
     'Elasticsearch\\Common\\Exceptions\\Serializer\\JsonErrorException' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/Serializer/JsonErrorException.php',
     'Elasticsearch\\Common\\Exceptions\\ServerErrorResponseException' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/ServerErrorResponseException.php',
     'Elasticsearch\\Common\\Exceptions\\TransportException' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/TransportException.php',
+    'Elasticsearch\\Common\\Exceptions\\Unauthorized401Exception' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/Unauthorized401Exception.php',
     'Elasticsearch\\Common\\Exceptions\\UnexpectedValueException' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/UnexpectedValueException.php',
     'Elasticsearch\\ConnectionPool\\AbstractConnectionPool' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/ConnectionPool/AbstractConnectionPool.php',
     'Elasticsearch\\ConnectionPool\\ConnectionPoolInterface' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/ConnectionPool/ConnectionPoolInterface.php',
@@ -412,6 +434,7 @@ return array(
     'Elasticsearch\\Endpoints\\Cluster\\Nodes\\Shutdown' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Nodes/Shutdown.php',
     'Elasticsearch\\Endpoints\\Cluster\\Nodes\\Stats' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Nodes/Stats.php',
     'Elasticsearch\\Endpoints\\Cluster\\PendingTasks' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/PendingTasks.php',
+    'Elasticsearch\\Endpoints\\Cluster\\RemoteInfo' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/RemoteInfo.php',
     'Elasticsearch\\Endpoints\\Cluster\\Reroute' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Reroute.php',
     'Elasticsearch\\Endpoints\\Cluster\\Settings\\Get' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Settings/Get.php',
     'Elasticsearch\\Endpoints\\Cluster\\Settings\\Put' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Settings/Put.php',
@@ -433,7 +456,6 @@ return array(
     'Elasticsearch\\Endpoints\\Indices\\Alias\\Exists' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Alias/Exists.php',
     'Elasticsearch\\Endpoints\\Indices\\Alias\\Get' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Alias/Get.php',
     'Elasticsearch\\Endpoints\\Indices\\Alias\\Put' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Alias/Put.php',
-    'Elasticsearch\\Endpoints\\Indices\\Aliases\\Get' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Aliases/Get.php',
     'Elasticsearch\\Endpoints\\Indices\\Aliases\\Update' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Aliases/Update.php',
     'Elasticsearch\\Endpoints\\Indices\\Analyze' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Analyze.php',
     'Elasticsearch\\Endpoints\\Indices\\Cache\\Clear' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Cache/Clear.php',
@@ -478,6 +500,7 @@ return array(
     'Elasticsearch\\Endpoints\\Info' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Info.php',
     'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\Delete' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Pipeline/Delete.php',
     'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\Get' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Pipeline/Get.php',
+    'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\ProcessorGrok' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/ProcessorGrok.php',
     'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\Put' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Pipeline/Put.php',
     'Elasticsearch\\Endpoints\\Ingest\\Simulate' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Simulate.php',
     'Elasticsearch\\Endpoints\\MPercolate' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/MPercolate.php',
@@ -513,7 +536,6 @@ return array(
     'Elasticsearch\\Endpoints\\Tasks\\TasksList' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Tasks/TasksList.php',
     'Elasticsearch\\Endpoints\\Template\\Delete' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Template/Delete.php',
     'Elasticsearch\\Endpoints\\Template\\Get' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Template/Get.php',
-    'Elasticsearch\\Endpoints\\Template\\Put' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Template/Put.php',
     'Elasticsearch\\Endpoints\\TermVectors' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/TermVectors.php',
     'Elasticsearch\\Endpoints\\Update' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Update.php',
     'Elasticsearch\\Endpoints\\UpdateByQuery' => $vendorDir . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/UpdateByQuery.php',
@@ -541,6 +563,46 @@ return array(
     'Firebase\\JWT\\SignatureInvalidException' => $vendorDir . '/firebase/php-jwt/src/SignatureInvalidException.php',
     'GDText\\Box' => $vendorDir . '/stil/gd-text/src/Box.php',
     'GDText\\Color' => $vendorDir . '/stil/gd-text/src/Color.php',
+    'GuzzleHttp\\Ring\\Client\\ClientUtils' => $vendorDir . '/guzzlehttp/ringphp/src/Client/ClientUtils.php',
+    'GuzzleHttp\\Ring\\Client\\CurlFactory' => $vendorDir . '/guzzlehttp/ringphp/src/Client/CurlFactory.php',
+    'GuzzleHttp\\Ring\\Client\\CurlHandler' => $vendorDir . '/guzzlehttp/ringphp/src/Client/CurlHandler.php',
+    'GuzzleHttp\\Ring\\Client\\CurlMultiHandler' => $vendorDir . '/guzzlehttp/ringphp/src/Client/CurlMultiHandler.php',
+    'GuzzleHttp\\Ring\\Client\\Middleware' => $vendorDir . '/guzzlehttp/ringphp/src/Client/Middleware.php',
+    'GuzzleHttp\\Ring\\Client\\MockHandler' => $vendorDir . '/guzzlehttp/ringphp/src/Client/MockHandler.php',
+    'GuzzleHttp\\Ring\\Client\\StreamHandler' => $vendorDir . '/guzzlehttp/ringphp/src/Client/StreamHandler.php',
+    'GuzzleHttp\\Ring\\Core' => $vendorDir . '/guzzlehttp/ringphp/src/Core.php',
+    'GuzzleHttp\\Ring\\Exception\\CancelledException' => $vendorDir . '/guzzlehttp/ringphp/src/Exception/CancelledException.php',
+    'GuzzleHttp\\Ring\\Exception\\CancelledFutureAccessException' => $vendorDir . '/guzzlehttp/ringphp/src/Exception/CancelledFutureAccessException.php',
+    'GuzzleHttp\\Ring\\Exception\\ConnectException' => $vendorDir . '/guzzlehttp/ringphp/src/Exception/ConnectException.php',
+    'GuzzleHttp\\Ring\\Exception\\RingException' => $vendorDir . '/guzzlehttp/ringphp/src/Exception/RingException.php',
+    'GuzzleHttp\\Ring\\Future\\BaseFutureTrait' => $vendorDir . '/guzzlehttp/ringphp/src/Future/BaseFutureTrait.php',
+    'GuzzleHttp\\Ring\\Future\\CompletedFutureArray' => $vendorDir . '/guzzlehttp/ringphp/src/Future/CompletedFutureArray.php',
+    'GuzzleHttp\\Ring\\Future\\CompletedFutureValue' => $vendorDir . '/guzzlehttp/ringphp/src/Future/CompletedFutureValue.php',
+    'GuzzleHttp\\Ring\\Future\\FutureArray' => $vendorDir . '/guzzlehttp/ringphp/src/Future/FutureArray.php',
+    'GuzzleHttp\\Ring\\Future\\FutureArrayInterface' => $vendorDir . '/guzzlehttp/ringphp/src/Future/FutureArrayInterface.php',
+    'GuzzleHttp\\Ring\\Future\\FutureInterface' => $vendorDir . '/guzzlehttp/ringphp/src/Future/FutureInterface.php',
+    'GuzzleHttp\\Ring\\Future\\FutureValue' => $vendorDir . '/guzzlehttp/ringphp/src/Future/FutureValue.php',
+    'GuzzleHttp\\Ring\\Future\\MagicFutureTrait' => $vendorDir . '/guzzlehttp/ringphp/src/Future/MagicFutureTrait.php',
+    'GuzzleHttp\\Stream\\AppendStream' => $vendorDir . '/guzzlehttp/streams/src/AppendStream.php',
+    'GuzzleHttp\\Stream\\AsyncReadStream' => $vendorDir . '/guzzlehttp/streams/src/AsyncReadStream.php',
+    'GuzzleHttp\\Stream\\BufferStream' => $vendorDir . '/guzzlehttp/streams/src/BufferStream.php',
+    'GuzzleHttp\\Stream\\CachingStream' => $vendorDir . '/guzzlehttp/streams/src/CachingStream.php',
+    'GuzzleHttp\\Stream\\DroppingStream' => $vendorDir . '/guzzlehttp/streams/src/DroppingStream.php',
+    'GuzzleHttp\\Stream\\Exception\\CannotAttachException' => $vendorDir . '/guzzlehttp/streams/src/Exception/CannotAttachException.php',
+    'GuzzleHttp\\Stream\\Exception\\SeekException' => $vendorDir . '/guzzlehttp/streams/src/Exception/SeekException.php',
+    'GuzzleHttp\\Stream\\FnStream' => $vendorDir . '/guzzlehttp/streams/src/FnStream.php',
+    'GuzzleHttp\\Stream\\GuzzleStreamWrapper' => $vendorDir . '/guzzlehttp/streams/src/GuzzleStreamWrapper.php',
+    'GuzzleHttp\\Stream\\InflateStream' => $vendorDir . '/guzzlehttp/streams/src/InflateStream.php',
+    'GuzzleHttp\\Stream\\LazyOpenStream' => $vendorDir . '/guzzlehttp/streams/src/LazyOpenStream.php',
+    'GuzzleHttp\\Stream\\LimitStream' => $vendorDir . '/guzzlehttp/streams/src/LimitStream.php',
+    'GuzzleHttp\\Stream\\MetadataStreamInterface' => $vendorDir . '/guzzlehttp/streams/src/MetadataStreamInterface.php',
+    'GuzzleHttp\\Stream\\NoSeekStream' => $vendorDir . '/guzzlehttp/streams/src/NoSeekStream.php',
+    'GuzzleHttp\\Stream\\NullStream' => $vendorDir . '/guzzlehttp/streams/src/NullStream.php',
+    'GuzzleHttp\\Stream\\PumpStream' => $vendorDir . '/guzzlehttp/streams/src/PumpStream.php',
+    'GuzzleHttp\\Stream\\Stream' => $vendorDir . '/guzzlehttp/streams/src/Stream.php',
+    'GuzzleHttp\\Stream\\StreamDecoratorTrait' => $vendorDir . '/guzzlehttp/streams/src/StreamDecoratorTrait.php',
+    'GuzzleHttp\\Stream\\StreamInterface' => $vendorDir . '/guzzlehttp/streams/src/StreamInterface.php',
+    'GuzzleHttp\\Stream\\Utils' => $vendorDir . '/guzzlehttp/streams/src/Utils.php',
     'Hashable' => $vendorDir . '/data-values/data-values/src/interfaces/Hashable.php',
     'HtmlFormatter\\HtmlFormatter' => $vendorDir . '/wikimedia/html-formatter/src/HtmlFormatter.php',
     'IPSet\\IPSet' => $vendorDir . '/wikimedia/ip-set/src/IPSet/IPSet.php',
@@ -1228,6 +1290,18 @@ return array(
     'Psy\\VersionUpdater\\GitHubChecker' => $vendorDir . '/psy/psysh/src/VersionUpdater/GitHubChecker.php',
     'Psy\\VersionUpdater\\IntervalChecker' => $vendorDir . '/psy/psysh/src/VersionUpdater/IntervalChecker.php',
     'Psy\\VersionUpdater\\NoopChecker' => $vendorDir . '/psy/psysh/src/VersionUpdater/NoopChecker.php',
+    'React\\Promise\\CancellablePromiseInterface' => $vendorDir . '/react/promise/src/CancellablePromiseInterface.php',
+    'React\\Promise\\CancellationQueue' => $vendorDir . '/react/promise/src/CancellationQueue.php',
+    'React\\Promise\\Deferred' => $vendorDir . '/react/promise/src/Deferred.php',
+    'React\\Promise\\Exception\\LengthException' => $vendorDir . '/react/promise/src/Exception/LengthException.php',
+    'React\\Promise\\ExtendedPromiseInterface' => $vendorDir . '/react/promise/src/ExtendedPromiseInterface.php',
+    'React\\Promise\\FulfilledPromise' => $vendorDir . '/react/promise/src/FulfilledPromise.php',
+    'React\\Promise\\LazyPromise' => $vendorDir . '/react/promise/src/LazyPromise.php',
+    'React\\Promise\\Promise' => $vendorDir . '/react/promise/src/Promise.php',
+    'React\\Promise\\PromiseInterface' => $vendorDir . '/react/promise/src/PromiseInterface.php',
+    'React\\Promise\\PromisorInterface' => $vendorDir . '/react/promise/src/PromisorInterface.php',
+    'React\\Promise\\RejectedPromise' => $vendorDir . '/react/promise/src/RejectedPromise.php',
+    'React\\Promise\\UnhandledRejectionException' => $vendorDir . '/react/promise/src/UnhandledRejectionException.php',
     'RemexHtml\\DOM\\DOMBuilder' => $vendorDir . '/wikimedia/remex-html/RemexHtml/DOM/DOMBuilder.php',
     'RemexHtml\\DOM\\DOMFormatter' => $vendorDir . '/wikimedia/remex-html/RemexHtml/DOM/DOMFormatter.php',
     'RemexHtml\\DOM\\DOMSerializer' => $vendorDir . '/wikimedia/remex-html/RemexHtml/DOM/DOMSerializer.php',
