@@ -11,6 +11,7 @@ class ComposerStaticInit_mediawiki_vendor
         'd1715cacc3c23b16a030645514266a76' => __DIR__ . '/..' . '/data-values/interfaces/Interfaces.php',
         '90559502573a0d473dc66fde5c0ff7e2' => __DIR__ . '/..' . '/data-values/common/Common.php',
         '0e6d7bf4a5811bfa5cf40c5ccd6fae6a' => __DIR__ . '/..' . '/symfony/polyfill-mbstring/bootstrap.php',
+        'ad155f8f1cf0d418fe49e248db8c661b' => __DIR__ . '/..' . '/react/promise/src/functions_include.php',
         'dc2c047038b9f5c71759f55f00aba99b' => __DIR__ . '/..' . '/wikibase/data-model/WikibaseDataModel.php',
         '5a494680c593293bd6035e42e2a6825c' => __DIR__ . '/..' . '/data-values/geo/Geo.php',
         'c15f8000750e4e04ba559d9745044427' => __DIR__ . '/..' . '/data-values/number/Number.php',
@@ -66,6 +67,7 @@ class ComposerStaticInit_mediawiki_vendor
         'R' => 
         array (
             'RemexHtml\\' => 10,
+            'React\\Promise\\' => 14,
         ),
         'P' => 
         array (
@@ -88,6 +90,8 @@ class ComposerStaticInit_mediawiki_vendor
         ),
         'G' => 
         array (
+            'GuzzleHttp\\Stream\\' => 18,
+            'GuzzleHttp\\Ring\\' => 16,
             'GDText\\' => 7,
         ),
         'F' => 
@@ -217,6 +221,10 @@ class ComposerStaticInit_mediawiki_vendor
         array (
             0 => __DIR__ . '/..' . '/wikimedia/remex-html/RemexHtml',
         ),
+        'React\\Promise\\' => 
+        array (
+            0 => __DIR__ . '/..' . '/react/promise/src',
+        ),
         'Psy\\' => 
         array (
             0 => __DIR__ . '/..' . '/psy/psysh/src',
@@ -244,6 +252,14 @@ class ComposerStaticInit_mediawiki_vendor
         'JsonSchema\\' => 
         array (
             0 => __DIR__ . '/..' . '/justinrainbow/json-schema/src/JsonSchema',
+        ),
+        'GuzzleHttp\\Stream\\' => 
+        array (
+            0 => __DIR__ . '/..' . '/guzzlehttp/streams/src',
+        ),
+        'GuzzleHttp\\Ring\\' => 
+        array (
+            0 => __DIR__ . '/..' . '/guzzlehttp/ringphp/src',
         ),
         'GDText\\' => 
         array (
@@ -542,6 +558,7 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Aggregation\\AbstractSimpleAggregation' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/AbstractSimpleAggregation.php',
         'Elastica\\Aggregation\\AbstractTermsAggregation' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/AbstractTermsAggregation.php',
         'Elastica\\Aggregation\\Avg' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/Avg.php',
+        'Elastica\\Aggregation\\AvgBucket' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/AvgBucket.php',
         'Elastica\\Aggregation\\BucketScript' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/BucketScript.php',
         'Elastica\\Aggregation\\Cardinality' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/Cardinality.php',
         'Elastica\\Aggregation\\Children' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/Children.php',
@@ -568,7 +585,9 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Aggregation\\SerialDiff' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/SerialDiff.php',
         'Elastica\\Aggregation\\SignificantTerms' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/SignificantTerms.php',
         'Elastica\\Aggregation\\Stats' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/Stats.php',
+        'Elastica\\Aggregation\\StatsBucket' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/StatsBucket.php',
         'Elastica\\Aggregation\\Sum' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/Sum.php',
+        'Elastica\\Aggregation\\SumBucket' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/SumBucket.php',
         'Elastica\\Aggregation\\Terms' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/Terms.php',
         'Elastica\\Aggregation\\TopHits' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/TopHits.php',
         'Elastica\\Aggregation\\ValueCount' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Aggregation/ValueCount.php',
@@ -629,6 +648,26 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Node\\Info' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Node/Info.php',
         'Elastica\\Node\\Stats' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Node/Stats.php',
         'Elastica\\Param' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Param.php',
+        'Elastica\\Pipeline' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Pipeline.php',
+        'Elastica\\Processor\\AbstractProcessor' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/AbstractProcessor.php',
+        'Elastica\\Processor\\Append' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Append.php',
+        'Elastica\\Processor\\Attachment' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Attachment.php',
+        'Elastica\\Processor\\Convert' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Convert.php',
+        'Elastica\\Processor\\Date' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Date.php',
+        'Elastica\\Processor\\DateIndexName' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/DateIndexName.php',
+        'Elastica\\Processor\\DotExpander' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/DotExpander.php',
+        'Elastica\\Processor\\Fail' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Fail.php',
+        'Elastica\\Processor\\Join' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Join.php',
+        'Elastica\\Processor\\Json' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Json.php',
+        'Elastica\\Processor\\Kv' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Kv.php',
+        'Elastica\\Processor\\Lowercase' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Lowercase.php',
+        'Elastica\\Processor\\Remove' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Remove.php',
+        'Elastica\\Processor\\Rename' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Rename.php',
+        'Elastica\\Processor\\Set' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Set.php',
+        'Elastica\\Processor\\Sort' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Sort.php',
+        'Elastica\\Processor\\Split' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Split.php',
+        'Elastica\\Processor\\Trim' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Trim.php',
+        'Elastica\\Processor\\Uppercase' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Processor/Uppercase.php',
         'Elastica\\Query' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query.php',
         'Elastica\\QueryBuilder' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/QueryBuilder.php',
         'Elastica\\QueryBuilder\\DSL' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/QueryBuilder/DSL.php',
@@ -643,7 +682,6 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Query\\AbstractGeoShape' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/AbstractGeoShape.php',
         'Elastica\\Query\\AbstractQuery' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/AbstractQuery.php',
         'Elastica\\Query\\AbstractSpanQuery' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/AbstractSpanQuery.php',
-        'Elastica\\Query\\Bool' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Bool.php',
         'Elastica\\Query\\BoolQuery' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/BoolQuery.php',
         'Elastica\\Query\\Boosting' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Boosting.php',
         'Elastica\\Query\\Common' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Common.php',
@@ -654,7 +692,6 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Query\\Fuzzy' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Fuzzy.php',
         'Elastica\\Query\\GeoBoundingBox' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/GeoBoundingBox.php',
         'Elastica\\Query\\GeoDistance' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/GeoDistance.php',
-        'Elastica\\Query\\GeoDistanceRange' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/GeoDistanceRange.php',
         'Elastica\\Query\\GeoPolygon' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/GeoPolygon.php',
         'Elastica\\Query\\GeoShapePreIndexed' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/GeoShapePreIndexed.php',
         'Elastica\\Query\\GeoShapeProvided' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/GeoShapeProvided.php',
@@ -662,7 +699,6 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Query\\HasChild' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/HasChild.php',
         'Elastica\\Query\\HasParent' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/HasParent.php',
         'Elastica\\Query\\Ids' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Ids.php',
-        'Elastica\\Query\\Indices' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Indices.php',
         'Elastica\\Query\\InnerHits' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/InnerHits.php',
         'Elastica\\Query\\Limit' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Limit.php',
         'Elastica\\Query\\Match' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Match.php',
@@ -682,11 +718,14 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Query\\Script' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Script.php',
         'Elastica\\Query\\Simple' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Simple.php',
         'Elastica\\Query\\SimpleQueryString' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SimpleQueryString.php',
+        'Elastica\\Query\\SpanContaining' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanContaining.php',
         'Elastica\\Query\\SpanFirst' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanFirst.php',
         'Elastica\\Query\\SpanMulti' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanMulti.php',
         'Elastica\\Query\\SpanNear' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanNear.php',
+        'Elastica\\Query\\SpanNot' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanNot.php',
         'Elastica\\Query\\SpanOr' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanOr.php',
         'Elastica\\Query\\SpanTerm' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanTerm.php',
+        'Elastica\\Query\\SpanWithin' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/SpanWithin.php',
         'Elastica\\Query\\Term' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Term.php',
         'Elastica\\Query\\Terms' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Terms.php',
         'Elastica\\Query\\Type' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Query/Type.php',
@@ -706,7 +745,6 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Script\\AbstractScript' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Script/AbstractScript.php',
         'Elastica\\Script\\Script' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Script/Script.php',
         'Elastica\\Script\\ScriptFields' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Script/ScriptFields.php',
-        'Elastica\\Script\\ScriptFile' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Script/ScriptFile.php',
         'Elastica\\Script\\ScriptId' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Script/ScriptId.php',
         'Elastica\\Scroll' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Scroll.php',
         'Elastica\\Search' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Search.php',
@@ -720,14 +758,13 @@ class ComposerStaticInit_mediawiki_vendor
         'Elastica\\Suggest\\Completion' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Suggest/Completion.php',
         'Elastica\\Suggest\\Phrase' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Suggest/Phrase.php',
         'Elastica\\Suggest\\Term' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Suggest/Term.php',
-        'Elastica\\Tool\\CrossIndex' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Tool/CrossIndex.php',
+        'Elastica\\Task' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Task.php',
         'Elastica\\Transport\\AbstractTransport' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/AbstractTransport.php',
         'Elastica\\Transport\\AwsAuthV4' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/AwsAuthV4.php',
         'Elastica\\Transport\\Guzzle' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/Guzzle.php',
         'Elastica\\Transport\\Http' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/Http.php',
         'Elastica\\Transport\\HttpAdapter' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/HttpAdapter.php',
         'Elastica\\Transport\\Https' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/Https.php',
-        'Elastica\\Transport\\Null' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/Null.php',
         'Elastica\\Transport\\NullTransport' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Transport/NullTransport.php',
         'Elastica\\Type' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Type.php',
         'Elastica\\Type\\AbstractType' => __DIR__ . '/..' . '/ruflin/elastica/lib/Elastica/Type/AbstractType.php',
@@ -759,6 +796,7 @@ class ComposerStaticInit_mediawiki_vendor
         'Elasticsearch\\Common\\Exceptions\\Serializer\\JsonErrorException' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/Serializer/JsonErrorException.php',
         'Elasticsearch\\Common\\Exceptions\\ServerErrorResponseException' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/ServerErrorResponseException.php',
         'Elasticsearch\\Common\\Exceptions\\TransportException' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/TransportException.php',
+        'Elasticsearch\\Common\\Exceptions\\Unauthorized401Exception' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/Unauthorized401Exception.php',
         'Elasticsearch\\Common\\Exceptions\\UnexpectedValueException' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Common/Exceptions/UnexpectedValueException.php',
         'Elasticsearch\\ConnectionPool\\AbstractConnectionPool' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/ConnectionPool/AbstractConnectionPool.php',
         'Elasticsearch\\ConnectionPool\\ConnectionPoolInterface' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/ConnectionPool/ConnectionPoolInterface.php',
@@ -806,6 +844,7 @@ class ComposerStaticInit_mediawiki_vendor
         'Elasticsearch\\Endpoints\\Cluster\\Nodes\\Shutdown' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Nodes/Shutdown.php',
         'Elasticsearch\\Endpoints\\Cluster\\Nodes\\Stats' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Nodes/Stats.php',
         'Elasticsearch\\Endpoints\\Cluster\\PendingTasks' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/PendingTasks.php',
+        'Elasticsearch\\Endpoints\\Cluster\\RemoteInfo' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/RemoteInfo.php',
         'Elasticsearch\\Endpoints\\Cluster\\Reroute' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Reroute.php',
         'Elasticsearch\\Endpoints\\Cluster\\Settings\\Get' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Settings/Get.php',
         'Elasticsearch\\Endpoints\\Cluster\\Settings\\Put' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Cluster/Settings/Put.php',
@@ -827,7 +866,6 @@ class ComposerStaticInit_mediawiki_vendor
         'Elasticsearch\\Endpoints\\Indices\\Alias\\Exists' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Alias/Exists.php',
         'Elasticsearch\\Endpoints\\Indices\\Alias\\Get' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Alias/Get.php',
         'Elasticsearch\\Endpoints\\Indices\\Alias\\Put' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Alias/Put.php',
-        'Elasticsearch\\Endpoints\\Indices\\Aliases\\Get' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Aliases/Get.php',
         'Elasticsearch\\Endpoints\\Indices\\Aliases\\Update' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Aliases/Update.php',
         'Elasticsearch\\Endpoints\\Indices\\Analyze' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Analyze.php',
         'Elasticsearch\\Endpoints\\Indices\\Cache\\Clear' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Indices/Cache/Clear.php',
@@ -872,6 +910,7 @@ class ComposerStaticInit_mediawiki_vendor
         'Elasticsearch\\Endpoints\\Info' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Info.php',
         'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\Delete' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Pipeline/Delete.php',
         'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\Get' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Pipeline/Get.php',
+        'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\ProcessorGrok' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/ProcessorGrok.php',
         'Elasticsearch\\Endpoints\\Ingest\\Pipeline\\Put' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Pipeline/Put.php',
         'Elasticsearch\\Endpoints\\Ingest\\Simulate' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Ingest/Simulate.php',
         'Elasticsearch\\Endpoints\\MPercolate' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/MPercolate.php',
@@ -907,7 +946,6 @@ class ComposerStaticInit_mediawiki_vendor
         'Elasticsearch\\Endpoints\\Tasks\\TasksList' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Tasks/TasksList.php',
         'Elasticsearch\\Endpoints\\Template\\Delete' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Template/Delete.php',
         'Elasticsearch\\Endpoints\\Template\\Get' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Template/Get.php',
-        'Elasticsearch\\Endpoints\\Template\\Put' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Template/Put.php',
         'Elasticsearch\\Endpoints\\TermVectors' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/TermVectors.php',
         'Elasticsearch\\Endpoints\\Update' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/Update.php',
         'Elasticsearch\\Endpoints\\UpdateByQuery' => __DIR__ . '/..' . '/elasticsearch/elasticsearch/src/Elasticsearch/Endpoints/UpdateByQuery.php',
@@ -935,6 +973,46 @@ class ComposerStaticInit_mediawiki_vendor
         'Firebase\\JWT\\SignatureInvalidException' => __DIR__ . '/..' . '/firebase/php-jwt/src/SignatureInvalidException.php',
         'GDText\\Box' => __DIR__ . '/..' . '/stil/gd-text/src/Box.php',
         'GDText\\Color' => __DIR__ . '/..' . '/stil/gd-text/src/Color.php',
+        'GuzzleHttp\\Ring\\Client\\ClientUtils' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Client/ClientUtils.php',
+        'GuzzleHttp\\Ring\\Client\\CurlFactory' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Client/CurlFactory.php',
+        'GuzzleHttp\\Ring\\Client\\CurlHandler' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Client/CurlHandler.php',
+        'GuzzleHttp\\Ring\\Client\\CurlMultiHandler' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Client/CurlMultiHandler.php',
+        'GuzzleHttp\\Ring\\Client\\Middleware' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Client/Middleware.php',
+        'GuzzleHttp\\Ring\\Client\\MockHandler' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Client/MockHandler.php',
+        'GuzzleHttp\\Ring\\Client\\StreamHandler' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Client/StreamHandler.php',
+        'GuzzleHttp\\Ring\\Core' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Core.php',
+        'GuzzleHttp\\Ring\\Exception\\CancelledException' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Exception/CancelledException.php',
+        'GuzzleHttp\\Ring\\Exception\\CancelledFutureAccessException' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Exception/CancelledFutureAccessException.php',
+        'GuzzleHttp\\Ring\\Exception\\ConnectException' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Exception/ConnectException.php',
+        'GuzzleHttp\\Ring\\Exception\\RingException' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Exception/RingException.php',
+        'GuzzleHttp\\Ring\\Future\\BaseFutureTrait' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/BaseFutureTrait.php',
+        'GuzzleHttp\\Ring\\Future\\CompletedFutureArray' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/CompletedFutureArray.php',
+        'GuzzleHttp\\Ring\\Future\\CompletedFutureValue' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/CompletedFutureValue.php',
+        'GuzzleHttp\\Ring\\Future\\FutureArray' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/FutureArray.php',
+        'GuzzleHttp\\Ring\\Future\\FutureArrayInterface' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/FutureArrayInterface.php',
+        'GuzzleHttp\\Ring\\Future\\FutureInterface' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/FutureInterface.php',
+        'GuzzleHttp\\Ring\\Future\\FutureValue' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/FutureValue.php',
+        'GuzzleHttp\\Ring\\Future\\MagicFutureTrait' => __DIR__ . '/..' . '/guzzlehttp/ringphp/src/Future/MagicFutureTrait.php',
+        'GuzzleHttp\\Stream\\AppendStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/AppendStream.php',
+        'GuzzleHttp\\Stream\\AsyncReadStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/AsyncReadStream.php',
+        'GuzzleHttp\\Stream\\BufferStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/BufferStream.php',
+        'GuzzleHttp\\Stream\\CachingStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/CachingStream.php',
+        'GuzzleHttp\\Stream\\DroppingStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/DroppingStream.php',
+        'GuzzleHttp\\Stream\\Exception\\CannotAttachException' => __DIR__ . '/..' . '/guzzlehttp/streams/src/Exception/CannotAttachException.php',
+        'GuzzleHttp\\Stream\\Exception\\SeekException' => __DIR__ . '/..' . '/guzzlehttp/streams/src/Exception/SeekException.php',
+        'GuzzleHttp\\Stream\\FnStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/FnStream.php',
+        'GuzzleHttp\\Stream\\GuzzleStreamWrapper' => __DIR__ . '/..' . '/guzzlehttp/streams/src/GuzzleStreamWrapper.php',
+        'GuzzleHttp\\Stream\\InflateStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/InflateStream.php',
+        'GuzzleHttp\\Stream\\LazyOpenStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/LazyOpenStream.php',
+        'GuzzleHttp\\Stream\\LimitStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/LimitStream.php',
+        'GuzzleHttp\\Stream\\MetadataStreamInterface' => __DIR__ . '/..' . '/guzzlehttp/streams/src/MetadataStreamInterface.php',
+        'GuzzleHttp\\Stream\\NoSeekStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/NoSeekStream.php',
+        'GuzzleHttp\\Stream\\NullStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/NullStream.php',
+        'GuzzleHttp\\Stream\\PumpStream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/PumpStream.php',
+        'GuzzleHttp\\Stream\\Stream' => __DIR__ . '/..' . '/guzzlehttp/streams/src/Stream.php',
+        'GuzzleHttp\\Stream\\StreamDecoratorTrait' => __DIR__ . '/..' . '/guzzlehttp/streams/src/StreamDecoratorTrait.php',
+        'GuzzleHttp\\Stream\\StreamInterface' => __DIR__ . '/..' . '/guzzlehttp/streams/src/StreamInterface.php',
+        'GuzzleHttp\\Stream\\Utils' => __DIR__ . '/..' . '/guzzlehttp/streams/src/Utils.php',
         'Hashable' => __DIR__ . '/..' . '/data-values/data-values/src/interfaces/Hashable.php',
         'HtmlFormatter\\HtmlFormatter' => __DIR__ . '/..' . '/wikimedia/html-formatter/src/HtmlFormatter.php',
         'IPSet\\IPSet' => __DIR__ . '/..' . '/wikimedia/ip-set/src/IPSet/IPSet.php',
@@ -1622,6 +1700,18 @@ class ComposerStaticInit_mediawiki_vendor
         'Psy\\VersionUpdater\\GitHubChecker' => __DIR__ . '/..' . '/psy/psysh/src/VersionUpdater/GitHubChecker.php',
         'Psy\\VersionUpdater\\IntervalChecker' => __DIR__ . '/..' . '/psy/psysh/src/VersionUpdater/IntervalChecker.php',
         'Psy\\VersionUpdater\\NoopChecker' => __DIR__ . '/..' . '/psy/psysh/src/VersionUpdater/NoopChecker.php',
+        'React\\Promise\\CancellablePromiseInterface' => __DIR__ . '/..' . '/react/promise/src/CancellablePromiseInterface.php',
+        'React\\Promise\\CancellationQueue' => __DIR__ . '/..' . '/react/promise/src/CancellationQueue.php',
+        'React\\Promise\\Deferred' => __DIR__ . '/..' . '/react/promise/src/Deferred.php',
+        'React\\Promise\\Exception\\LengthException' => __DIR__ . '/..' . '/react/promise/src/Exception/LengthException.php',
+        'React\\Promise\\ExtendedPromiseInterface' => __DIR__ . '/..' . '/react/promise/src/ExtendedPromiseInterface.php',
+        'React\\Promise\\FulfilledPromise' => __DIR__ . '/..' . '/react/promise/src/FulfilledPromise.php',
+        'React\\Promise\\LazyPromise' => __DIR__ . '/..' . '/react/promise/src/LazyPromise.php',
+        'React\\Promise\\Promise' => __DIR__ . '/..' . '/react/promise/src/Promise.php',
+        'React\\Promise\\PromiseInterface' => __DIR__ . '/..' . '/react/promise/src/PromiseInterface.php',
+        'React\\Promise\\PromisorInterface' => __DIR__ . '/..' . '/react/promise/src/PromisorInterface.php',
+        'React\\Promise\\RejectedPromise' => __DIR__ . '/..' . '/react/promise/src/RejectedPromise.php',
+        'React\\Promise\\UnhandledRejectionException' => __DIR__ . '/..' . '/react/promise/src/UnhandledRejectionException.php',
         'RemexHtml\\DOM\\DOMBuilder' => __DIR__ . '/..' . '/wikimedia/remex-html/RemexHtml/DOM/DOMBuilder.php',
         'RemexHtml\\DOM\\DOMFormatter' => __DIR__ . '/..' . '/wikimedia/remex-html/RemexHtml/DOM/DOMFormatter.php',
         'RemexHtml\\DOM\\DOMSerializer' => __DIR__ . '/..' . '/wikimedia/remex-html/RemexHtml/DOM/DOMSerializer.php',
