@@ -1270,7 +1270,6 @@ class Client
         /** @var \Elasticsearch\Endpoints\Script\Put $endpoint */
         $endpoint = $endpointBuilder('Script\Put');
         $endpoint->setID($id)
-                 ->setLang($lang)
                  ->setBody($body);
         $endpoint->setParams($params);
 
@@ -1317,30 +1316,6 @@ class Client
         $endpoint = $endpointBuilder('Template\Delete');
         $endpoint->setID($id);
         $endpoint->setParams($params);
-
-        return $this->performRequest($endpoint);
-    }
-
-    /**
-     * $params['id']   = (string) The search template ID (Required)
-     *
-     * @param $params array Associative array of parameters
-     *
-     * @return array
-     */
-    public function putTemplate($params)
-    {
-        $id   = $this->extractArgument($params, 'id');
-        $body = $this->extractArgument($params, 'body');
-
-        /** @var callback $endpointBuilder */
-        $endpointBuilder = $this->endpoints;
-
-        /** @var \Elasticsearch\Endpoints\Template\Put $endpoint */
-        $endpoint = $endpointBuilder('Template\Put');
-        $endpoint->setID($id)
-            ->setBody($body)
-            ->setParams($params);
 
         return $this->performRequest($endpoint);
     }
@@ -1564,7 +1539,7 @@ class Client
 
     /**
      * @param $endpoint AbstractEndpoint
-     * 
+     *
      * @throws \Exception
      * @return array
      */
