@@ -89,9 +89,6 @@ class ClientBuilder
     /** @var null|bool|string */
     private $sslVerification = null;
 
-    /** @var bool  */
-    private $allowBadJSON = false;
-
     /**
      * @return ClientBuilder
      */
@@ -165,8 +162,8 @@ class ClientBuilder
     }
 
     /**
-     * @param array $singleParams
      * @param array $multiParams
+     * @param array $singleParams
      * @throws \RuntimeException
      * @return callable
      */
@@ -217,6 +214,7 @@ class ClientBuilder
 
     /**
      * @param $path string
+     * @param int $level
      * @return \Monolog\Logger\Logger
      */
     public static function defaultLogger($path, $level = Logger::WARNING)
@@ -309,6 +307,10 @@ class ClientBuilder
      */
     public function setLogger($logger)
     {
+        if (!$logger instanceof LoggerInterface) {
+            throw new InvalidArgumentException('$logger must implement \Psr\Log\LoggerInterface!');
+        }
+
         $this->logger = $logger;
 
         return $this;
@@ -320,6 +322,10 @@ class ClientBuilder
      */
     public function setTracer($tracer)
     {
+        if (!$tracer instanceof LoggerInterface) {
+            throw new InvalidArgumentException('$tracer must implement \Psr\Log\LoggerInterface!');
+        }
+
         $this->tracer = $tracer;
 
         return $this;
@@ -428,24 +434,11 @@ class ClientBuilder
         return $this;
     }
 
-    public function allowBadJSONSerialization()
-    {
-        $this->allowBadJSON = true;
-        return $this;
-    }
-
     /**
      * @return Client
      */
     public function build()
     {
-        if(!defined('JSON_PRESERVE_ZERO_FRACTION') && $this->allowBadJSON === false) {
-            throw new RuntimeException("Your version of PHP / json-ext does not support the constant 'JSON_PRESERVE_ZERO_FRACTION',".
-            " which is important for proper type mapping in Elasticsearch. Please upgrade your PHP or json-ext.\n".
-            "If you are unable to upgrade, and are willing to accept the consequences, you may use the allowBadJSONSerialization()".
-            " method on the ClientBuilder to bypass this limitation.");
-        }
-
         $this->buildLoggers();
 
         if (is_null($this->handler)) {
@@ -489,16 +482,16 @@ class ClientBuilder
                 $this->connectionParams = [];
             }
 
-            // Make sure we are setting Content-type and Accept (unless the user has explicitly
+            // Make sure we are setting Content-Type and Accept (unless the user has explicitly
             // overridden it
             if (isset($this->connectionParams['client']['headers']) === false) {
                 $this->connectionParams['client']['headers'] = [
-                    'Content-type' => ['application/json'],
+                    'Content-Type' => ['application/json'],
                     'Accept' => ['application/json']
                 ];
             } else {
-                if (isset($this->connectionParams['client']['headers']['Content-type']) === false) {
-                    $this->connectionParams['client']['headers']['Content-type'] = ['application/json'];
+                if (isset($this->connectionParams['client']['headers']['Content-Type']) === false) {
+                    $this->connectionParams['client']['headers']['Content-Type'] = ['application/json'];
                 }
                 if (isset($this->connectionParams['client']['headers']['Accept']) === false) {
                     $this->connectionParams['client']['headers']['Accept'] = ['application/json'];
@@ -525,7 +518,7 @@ class ClientBuilder
 
             $this->endpoint = function ($class) use ($serializer) {
                 $fullPath = '\\Elasticsearch\\Endpoints\\' . $class;
-                if ($class === 'Bulk' || $class === 'Msearch' || $class === 'MPercolate') {
+                if ($class === 'Bulk' || $class === 'Msearch' || $class === 'MsearchTemplate' || $class === 'MPercolate') {
                     return new $fullPath($serializer);
                 } else {
                     return new $fullPath();
@@ -573,13 +566,15 @@ class ClientBuilder
                 $connections,
                 $this->selector,
                 $this->connectionFactory,
-                $this->connectionPoolArgs);
+                $this->connectionPoolArgs
+            );
         } elseif (is_null($this->connectionPool)) {
             $this->connectionPool = new StaticNoPingConnectionPool(
                 $connections,
                 $this->selector,
                 $this->connectionFactory,
-                $this->connectionPoolArgs);
+                $this->connectionPoolArgs
+            );
         }
 
         if (is_null($this->retries)) {
@@ -628,7 +623,7 @@ class ClientBuilder
             if (is_string($host)) {
                 $host = $this->prependMissingScheme($host);
                 $host = $this->extractURIParts($host);
-            } else if (is_array($host)) {
+            } elseif (is_array($host)) {
                 $host = $this->normalizeExtendedHost($host);
             } else {
                 $this->logger->error("Could not parse host: ".print_r($host, true));
@@ -644,7 +639,8 @@ class ClientBuilder
      * @param $host
      * @return array
      */
-    private function normalizeExtendedHost($host) {
+    private function normalizeExtendedHost($host)
+    {
         if (isset($host['host']) === false) {
             $this->logger->error("Required 'host' was not defined in extended format: ".print_r($host, true));
             throw new RuntimeException("Required 'host' was not defined in extended format: ".print_r($host, true));
