@@ -16,7 +16,7 @@ interface SelectorInterface
     /**
      * Perform logic to select a single ConnectionInterface instance from the array provided
      *
-     * @param  ConnectionInterface[] $connections an array of ConnectionInterface instances to choose from
+     * @param  \Elasticsearch\Connections\ConnectionInterface[] $connections an array of ConnectionInterface instances to choose from
      *
      * @return \Elasticsearch\Connections\ConnectionInterface
      */
