@@ -39,7 +39,7 @@ class FieldCaps extends AbstractEndpoint
     {
         $index = $this->index;
 
-        if (isset($index) === true ) {
+        if (isset($index) === true) {
             return "/$index/_field_caps";
         } else {
             return "/_field_caps";
