@@ -16,25 +16,6 @@ use Elasticsearch\Common\Exceptions;
  */
 class Put extends AbstractEndpoint
 {
-    /** @var  String */
-    private $lang;
-
-    /**
-     * @param $lang
-     *
-     * @return $this
-     */
-    public function setLang($lang)
-    {
-        if (isset($lang) !== true) {
-            return $this;
-        }
-
-        $this->lang = $lang;
-
-        return $this;
-    }
-
     /**
      * @param array $body
      *
@@ -57,19 +38,13 @@ class Put extends AbstractEndpoint
      */
     public function getURI()
     {
-        if (isset($this->lang) !== true) {
-            throw new Exceptions\RuntimeException(
-                'lang is required for Put'
-            );
-        }
         if (isset($this->id) !== true) {
             throw new Exceptions\RuntimeException(
                 'id is required for put'
             );
         }
         $id   = $this->id;
-        $lang = $this->lang;
-        $uri  = "/_scripts/$lang/$id";
+        $uri  = "/_scripts/$id";
 
         return $uri;
     }
