@@ -98,7 +98,8 @@ class Search extends AbstractEndpoint
             'terminate_after',
             'stored_fields',
             'batched_reduce_size',
-            'typed_keys'
+            'typed_keys',
+            'pre_filter_shard_size'
         );
     }
 
