@@ -88,7 +88,8 @@ class Get extends AbstractEndpoint
     {
         return array(
             'master_timeout',
-            'ignore_unavailable'
+            'ignore_unavailable',
+            'verbose'
         );
     }
 
