@@ -2,6 +2,7 @@
 
 namespace Elasticsearch\Serializers;
 
+use Elasticsearch\Common\Exceptions;
 use Elasticsearch\Common\Exceptions\Serializer\JsonErrorException;
 
 /**
@@ -15,13 +16,6 @@ use Elasticsearch\Common\Exceptions\Serializer\JsonErrorException;
  */
 class SmartSerializer implements SerializerInterface
 {
-    private $PHP_VERSION;
-
-    public function __construct()
-    {
-        $this->PHP_VERSION = phpversion();
-    }
-
     /**
      * Serialize assoc array into JSON string
      *
@@ -34,10 +28,9 @@ class SmartSerializer implements SerializerInterface
         if (is_string($data) === true) {
             return $data;
         } else {
-            if (version_compare($this->PHP_VERSION, '5.6.6', '<') || ! defined('JSON_PRESERVE_ZERO_FRACTION')) {
-                $data = json_encode($data);
-            } else {
-                $data = json_encode($data, JSON_PRESERVE_ZERO_FRACTION);
+            $data = json_encode($data, JSON_PRESERVE_ZERO_FRACTION);
+            if ($data === false) {
+                throw new Exceptions\RuntimeException("Failed to JSON encode: ".json_last_error());
             }
             if ($data === '[]') {
                 return '{}';
