@@ -10,7 +10,6 @@ use Elastica\Exception\InvalidException;
  */
 class GeoDistance extends AbstractAggregation
 {
-    const DISTANCE_TYPE_SLOPPY_ARC = 'sloppy_arc';
     const DISTANCE_TYPE_ARC = 'arc';
     const DISTANCE_TYPE_PLANE = 'plane';
 
@@ -93,7 +92,7 @@ class GeoDistance extends AbstractAggregation
     /**
      * Set the method by which distances will be calculated.
      *
-     * @param string $distanceType see DISTANCE_TYPE_* constants for options. Defaults to sloppy_arc.
+     * @param string $distanceType see DISTANCE_TYPE_* constants for options. Defaults to arc.
      *
      * @return $this
      */
