@@ -263,9 +263,9 @@ class Settings
      * Sets the index refresh interval.
      *
      * Value can be for example 3s for 3 seconds or
-     * 5m for 5 minutes. -1 refreshing is disabled.
+     * 5m for 5 minutes. -1 to disabled refresh.
      *
-     * @param int $interval Number of milliseconds
+     * @param string $interval Duration of the refresh interval
      *
      * @return \Elastica\Response Response object
      */
