@@ -47,7 +47,9 @@ class MultiBuilder implements MultiBuilderInterface
         reset($searches);
 
         foreach ($data['responses'] as $responseData) {
-            list($key, $search) = each($searches);
+            $search = current($searches);
+            $key = key($searches);
+            next($searches);
 
             $resultSets[$key] = $this->buildResultSet(new Response($responseData), $search);
         }
