@@ -12,11 +12,6 @@ class GeoDistance extends AbstractGeoDistance
 {
     const DISTANCE_TYPE_ARC = 'arc';
     const DISTANCE_TYPE_PLANE = 'plane';
-    const DISTANCE_TYPE_SLOPPY_ARC = 'sloppy_arc';
-
-    const OPTIMIZE_BBOX_MEMORY = 'memory';
-    const OPTIMIZE_BBOX_INDEXED = 'indexed';
-    const OPTIMIZE_BBOX_NONE = 'none';
 
     /**
      * Create GeoDistance object.
@@ -49,27 +44,13 @@ class GeoDistance extends AbstractGeoDistance
     /**
      * See DISTANCE_TYPE_* constants.
      *
-     * @param string $distanceType
+     * @param string $distanceType, default arc
      *
      * @return $this
      */
     public function setDistanceType($distanceType)
     {
         $this->setParam('distance_type', $distanceType);
-
-        return $this;
-    }
-
-    /**
-     * See OPTIMIZE_BBOX_* constants.
-     *
-     * @param string $optimizeBbox
-     *
-     * @return $this
-     */
-    public function setOptimizeBbox($optimizeBbox)
-    {
-        $this->setParam('optimize_bbox', $optimizeBbox);
 
         return $this;
     }
