@@ -145,10 +145,8 @@ class Mapping
      * _id
      * _type
      * _source
-     * _all
      * _analyzer
      * _boost
-     * _parent
      * _routing
      * _index
      * _size
@@ -177,43 +175,7 @@ class Mapping
      */
     public function getParam($key)
     {
-        return isset($this->_mapping[$key]) ? $this->_mapping[$key] : null;
-    }
-
-    /**
-     * Sets params for the "_all" field.
-     *
-     * @param array $params _all Params (enabled, store, term_vector, analyzer)
-     *
-     * @return $this
-     */
-    public function setAllField(array $params)
-    {
-        return $this->setParam('_all', $params);
-    }
-
-    /**
-     * Enables the "_all" field.
-     *
-     * @param bool $enabled OPTIONAL (default = true)
-     *
-     * @return $this
-     */
-    public function enableAllField($enabled = true)
-    {
-        return $this->setAllField(['enabled' => $enabled]);
-    }
-
-    /**
-     * Set parent type.
-     *
-     * @param string $type Parent type
-     *
-     * @return $this
-     */
-    public function setParent($type)
-    {
-        return $this->setParam('_parent', ['type' => $type]);
+        return $this->_mapping[$key] ?? null;
     }
 
     /**
